refator:테스트를 위한 play  기능의 분리
