+++ src/index.js
@@ -1,7 +1,7 @@
 import { createRandomNumber } from './computeNumbers.js';
 import { checkInputs } from './inputCheck.js';
 import { showResult } from './utils.js';
-import { checkNothing, checkStrike } from './compareNumber.js';
+import { play } from './play.js';
 
 export default class BaseballGame {
   constructor() {
@@ -26,23 +26,6 @@ export default class BaseballGame {
     return randomNum;
   }
 
-  play(computerInputNumbers, userInputNumbers) {
-    const BALL_STRIKE_COUNT = checkNothing(userInputNumbers, computerInputNumbers);
-    const STRIKE_COUNT = checkStrike(userInputNumbers, computerInputNumbers);
-    const BALL_COUNT = BALL_STRIKE_COUNT - STRIKE_COUNT;
-    if (BALL_STRIKE_COUNT === 0) {
-      return '낫싱';
-    }
-    if (STRIKE_COUNT === 0) {
-      return `${BALL_COUNT}볼`;
-    }
-    if (STRIKE_COUNT === 3) {
-      this.restartArticle.style.display = 'block';
-      return '<strong>🎉 정답을 맞추셨습니다! 🎉<strong>';
-    }
-    return `${BALL_COUNT}볼 ${STRIKE_COUNT}스트라이크`;
-  }
-
   pressSubmitButton() {
     this.submitButton.onclick = (e) => {
       e.preventDefault();
@@ -51,7 +34,7 @@ export default class BaseballGame {
         alert('값을 확인해주세요!');
         return;
       }
-      const RESULT = this.play(userInputValue, this.computerInputNumbers);
+      const RESULT = play(userInputValue, this.computerInputNumbers);
       showResult(RESULT);
     };
   }
